Add chalice work flow, convert to chalice endpoint and move pathfinding class to chalicelib directory
